requirements.txt
python-dotenv==0.20.0
pika==1.3.2
numpy==1.24.4
opencv-python==4.8.0.74
